+++ src/geos.h
@@ -52,7 +52,7 @@ enum {
   PGERR_GEOMETRY_TYPE,
   PGERR_MULTIPOINT_WITH_POINT_EMPTY,
   PGERR_EMPTY_GEOMETRY,
-  PGWARN_GEOS_EXCEPTION  // raise the GEOS error as a warning instead of exception
+  PGWARN_INVALID_WKB  // raise the GEOS WKB error as a warning instead of exception
 };
 
 // Define how the states are handled by CPython
@@ -86,8 +86,9 @@ enum {
     case PGERR_EMPTY_GEOMETRY:                                                           \
       PyErr_SetString(PyExc_ValueError, "One of the Geometry inputs is empty.");         \
       break;                                                                             \
-    case PGWARN_GEOS_EXCEPTION:                                                          \
-      PyErr_WarnEx(PyExc_Warning, last_error, 0);                                        \
+    case PGWARN_INVALID_WKB:                                                             \
+      PyErr_WarnFormat(PyExc_Warning, 0,                                                 \
+                       "Invalid WKB: geometry is returned as None. %s", last_error);     \
       break;                                                                             \
     default:                                                                             \
       PyErr_Format(PyExc_RuntimeError,                                                   \